EDU-293 Split DB schemas for lessons and documents sections
