+++ src/domain/schemas/shared-schemas.js
@@ -5,21 +5,9 @@ export const boolStringSchema = joi.any().valid('true', 'false', true, false);
 
 export const slugSchema = joi.string().allow('').required();
 
-export const sectionDBSchema = joi.object({
-  revision: idOrKeySchema.required(),
-  key: idOrKeySchema.required(),
-  deletedOn: joi.date().allow(null).required(),
-  deletedBy: idOrKeySchema.allow(null).required(),
-  deletedBecause: joi.string().allow(null).required(),
-  type: joi.string().required(),
-  content: joi.alternatives().try(
-    joi.object().required(),
-    joi.any().allow(null).required()
-  ).required()
-});
-
 export const sectionSchema = joi.object({
   key: idOrKeySchema.required(),
+  revision: idOrKeySchema.allow(null),
   type: joi.string().required(),
   content: joi.alternatives().try(
     joi.object().required(),
